Differentiate between uninstall and fulluninstall.
The 'uninstall' profile belongs to the 'base' profile, which is the
one that gets applied when you install ATContentTypes.

The 'fulluninstall' profile belongs to the 'default' profile, which
can only be installed through portal_setup.

Do not try to install p.a.imaging/widgets.  Their default profiles are
dummies in Plone 5.